+++ packages/graph-explorer/src/connector/queries.ts
@@ -1,17 +1,18 @@
 import { QueryClient, queryOptions } from "@tanstack/react-query";
 import {
-  EdgeDetailsRequest,
   EdgeDetailsResponse,
-  Explorer,
   KeywordSearchRequest,
-  KeywordSearchResponse,
   SchemaResponse,
-  toMappedQueryResults,
-  VertexDetailsRequest,
   VertexDetailsResponse,
 } from "./useGEFetchTypes";
-import { Edge, Vertex, VertexId } from "@/core";
-import { updateSchemaPrefixes } from "@/core/StateProvider/schema";
+import { Edge, EdgeId, Vertex, VertexId } from "@/core";
+import {
+  UpdateSchemaHandler,
+  updateSchemaPrefixes,
+} from "@/core/StateProvider/schema";
+import { logger } from "@/utils";
+import { emptyExplorer } from "./emptyExplorer";
+import { GraphExplorerMeta } from "@/core/queryClient";
 
 /**
  * Fetches the schema from the given explorer and updates the local cache with the new schema.
@@ -19,12 +20,12 @@ import { updateSchemaPrefixes } from "@/core/StateProvider/schema";
  * @param explorer The explorer to use for fetching the schema.
  */
 export function schemaSyncQuery(
-  updateLocalCache: (schema: SchemaResponse) => void,
-  explorer: Explorer
+  updateLocalCache: (schema: SchemaResponse) => void
 ) {
   return queryOptions({
-    queryKey: ["schema", explorer],
-    queryFn: async ({ signal }) => {
+    queryKey: ["schema"],
+    queryFn: async ({ signal, meta }) => {
+      const explorer = getExplorer(meta);
       let schema = await explorer.fetchSchema({ signal });
 
       // Update the prefixes for sparql connections
@@ -47,19 +48,15 @@ export function schemaSyncQuery(
  */
 export function searchQuery(
   request: KeywordSearchRequest,
-  updateSchema: (entities: { vertices: Vertex[]; edges: Edge[] }) => void,
-  explorer: Explorer,
-  queryClient: QueryClient
+  updateSchema: UpdateSchemaHandler
 ) {
   return queryOptions({
-    queryKey: ["keyword-search", request, explorer, queryClient],
-    queryFn: async ({ signal }): Promise<KeywordSearchResponse> => {
-      if (!request) {
-        return toMappedQueryResults({});
-      }
+    queryKey: ["keyword-search", request],
+    queryFn: async ({ signal, meta, client }) => {
+      const explorer = getExplorer(meta);
       const results = await explorer.keywordSearch(request, { signal });
 
-      updateVertexDetailsCache(explorer, queryClient, results.vertices);
+      updateVertexDetailsCache(client, results.vertices);
       updateSchema(results);
 
       return results;
@@ -84,16 +81,17 @@ export type NeighborCountsQueryResponse = {
  * @param explorer The service client to use for fetching the neighbors count.
  * @returns The count of neighbors for the given node as a total and per type.
  */
-export function neighborsCountQuery(
-  request: NeighborCountsQueryRequest,
-  explorer: Explorer
-) {
+export function neighborsCountQuery(request: NeighborCountsQueryRequest) {
   return queryOptions({
-    queryKey: ["neighborsCount", request, explorer],
-    queryFn: async (): Promise<NeighborCountsQueryResponse> => {
-      const result = await explorer.fetchNeighborsCount({
-        vertexId: request.vertexId,
-      });
+    queryKey: ["neighborsCount", request],
+    queryFn: async ({ signal, meta }) => {
+      const explorer = getExplorer(meta);
+      const result = await explorer.fetchNeighborsCount(
+        {
+          vertexId: request.vertexId,
+        },
+        { signal }
+      );
 
       return {
         nodeId: request.vertexId,
@@ -110,72 +108,78 @@ export function neighborsCountQuery(
  * @param explorer The service client to use for fetching.
  * @returns An object with the total nodes for the given node type.
  */
-export const nodeCountByNodeTypeQuery = (
-  nodeType: string,
-  explorer: Explorer
-) =>
-  queryOptions({
-    queryKey: ["node-count-by-node-type", nodeType, explorer],
-    queryFn: () =>
-      explorer.fetchVertexCountsByType({
-        label: nodeType,
-      }),
-  });
-
-export function vertexDetailsQuery(
-  request: VertexDetailsRequest,
-  explorer: Explorer
-) {
+export function nodeCountByNodeTypeQuery(nodeType: string) {
   return queryOptions({
-    queryKey: ["db", "vertex", "details", request, explorer],
-    queryFn: ({ signal }): Promise<VertexDetailsResponse> =>
-      explorer.vertexDetails(request, { signal }),
+    queryKey: ["node-count-by-node-type", nodeType],
+    queryFn: ({ signal, meta }) => {
+      const explorer = getExplorer(meta);
+      return explorer.fetchVertexCountsByType(
+        {
+          label: nodeType,
+        },
+        { signal }
+      );
+    },
   });
 }
 
-export function edgeDetailsQuery(
-  request: EdgeDetailsRequest,
-  explorer: Explorer
-) {
+export function vertexDetailsQuery(vertexId: VertexId) {
+  const request = { vertexId };
+
   return queryOptions({
-    queryKey: ["db", "edge", "details", request, explorer],
-    queryFn: ({ signal }): Promise<EdgeDetailsResponse> =>
-      explorer.edgeDetails(request, { signal }),
+    queryKey: ["db", "vertex", "details", request],
+    queryFn: ({ signal, meta }) => {
+      const explorer = getExplorer(meta);
+      return explorer.vertexDetails(request, { signal });
+    },
+  });
+}
+
+export function edgeDetailsQuery(edgeId: EdgeId) {
+  const request = { edgeId };
+
+  return queryOptions({
+    queryKey: ["db", "edge", "details", request],
+    queryFn: ({ signal, meta }) => {
+      const explorer = getExplorer(meta);
+      return explorer.edgeDetails(request, { signal });
+    },
   });
 }
 
 /** Sets the vertex details cache for the given vertices. */
 export function updateVertexDetailsCache(
-  explorer: Explorer,
   queryClient: QueryClient,
   vertices: Vertex[]
 ) {
   for (const vertex of vertices.filter(v => !v.__isFragment)) {
-    const request: VertexDetailsRequest = {
-      vertexId: vertex.id,
-    };
     const response: VertexDetailsResponse = {
       vertex,
     };
-    const queryKey = vertexDetailsQuery(request, explorer).queryKey;
+    const queryKey = vertexDetailsQuery(vertex.id).queryKey;
     queryClient.setQueryData(queryKey, response);
   }
 }
 
 /** Sets the edge details cache for the given edges. */
 export function updateEdgeDetailsCache(
-  explorer: Explorer,
   queryClient: QueryClient,
   edges: Edge[]
 ) {
   for (const edge of edges.filter(e => !e.__isFragment)) {
-    const request: EdgeDetailsRequest = {
-      edgeId: edge.id,
-    };
     const response: EdgeDetailsResponse = {
       edge,
     };
-    const queryKey = edgeDetailsQuery(request, explorer).queryKey;
+    const queryKey = edgeDetailsQuery(edge.id).queryKey;
     queryClient.setQueryData(queryKey, response);
   }
+}
+
+/** Extracts the explorer from the meta objects */
+function getExplorer(meta: GraphExplorerMeta | undefined) {
+  if (!meta?.explorer) {
+    logger.warn("No explorer found in the query client meta object");
+    return emptyExplorer;
+  }
+  return meta.explorer;
 }