+++ packages/graph-explorer/src/core/ConnectedProvider/ConnectedProvider.tsx
@@ -1,9 +1,5 @@
 import { type PropsWithChildren } from "react";
-import {
-  QueryCache,
-  QueryClient,
-  QueryClientProvider,
-} from "@tanstack/react-query";
+import { QueryClientProvider } from "@tanstack/react-query";
 import { NotificationProvider } from "@/components/NotificationProvider";
 import Toast from "@/components/Toast";
 import AppStatusLoader from "@/core/AppStatusLoader";
@@ -14,47 +10,11 @@ import { emotionTransform, MantineEmotionProvider } from "@mantine/emotion";
 import { ErrorBoundary } from "react-error-boundary";
 import AppErrorPage from "@/core/AppErrorPage";
 import { TooltipProvider } from "@/components";
-import { logger, NetworkError } from "@/utils";
-
-function exponentialBackoff(attempt: number): number {
-  return Math.min(attempt > 1 ? 2 ** attempt * 1000 : 1000, 30 * 1000);
-}
-
-const MAX_RETRIES = 3;
-const HTTP_STATUS_TO_NOT_RETRY = [400, 401, 403, 404, 429];
-
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      retry: (failureCount, error) => {
-        if (failureCount >= MAX_RETRIES) {
-          return false;
-        }
-        if (
-          error instanceof NetworkError &&
-          HTTP_STATUS_TO_NOT_RETRY.includes(error.statusCode)
-        ) {
-          logger.debug(
-            "Aborting retry due to HTTP status code:",
-            error.statusCode
-          );
-          return false;
-        }
-        return true;
-      },
-      retryDelay: exponentialBackoff,
-      staleTime: 1000 * 60, // 1 minute cache
-      refetchOnWindowFocus: false,
-    },
-  },
-  queryCache: new QueryCache({
-    onError(error, query) {
-      logger.error("Query failed to execute:", query.queryKey, error);
-    },
-  }),
-});
+import { useAtomValue } from "jotai";
+import { queryClientAtom } from "../connector";
 
 export default function ConnectedProvider({ children }: PropsWithChildren) {
+  const queryClient = useAtomValue(queryClientAtom);
   return (
     <ErrorBoundary FallbackComponent={AppErrorPage}>
       <QueryClientProvider client={queryClient}>