+++ packages/graph-explorer/src/hooks/useEdgeDetailsQuery.ts
@@ -1,9 +1,0 @@
-import { edgeDetailsQuery } from "@/connector";
-import { EdgeId, useExplorer } from "@/core";
-import { useQuery } from "@tanstack/react-query";
-
-export function useEdgeDetailsQuery(edgeId: EdgeId) {
-  const explorer = useExplorer();
-  const query = useQuery(edgeDetailsQuery({ edgeId }, explorer));
-  return query;
-}