+++ packages/graph-explorer/src/hooks/useMaterializeEdges.ts
@@ -1,11 +1,10 @@
-import { EdgeDetailsRequest, edgeDetailsQuery } from "@/connector";
-import { useExplorer, Edge, toEdgeMap, EdgeId } from "@/core";
+import { edgeDetailsQuery } from "@/connector";
+import { Edge, toEdgeMap, EdgeId } from "@/core";
 import { useQueryClient } from "@tanstack/react-query";
 
 /** Fetch the details if the edge is a fragment */
 export function useMaterializeEdges() {
   const queryClient = useQueryClient();
-  const explorer = useExplorer();
 
   return async (edges: Map<EdgeId, Edge>) => {
     const responses = await Promise.all(
@@ -14,11 +13,8 @@ export function useMaterializeEdges() {
           return edge;
         }
 
-        const request: EdgeDetailsRequest = {
-          edgeId: edge.id,
-        };
         const response = await queryClient.ensureQueryData(
-          edgeDetailsQuery(request, explorer)
+          edgeDetailsQuery(edge.id)
         );
         return response.edge;
       })